+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [Unreleased]
+
+* Add `NotificationClient.SendPrecompiledLetter` method.
+* Fixed `NotificationResponse.Equals` and `LetterNotificationResponse.Equals` for instances with `.template` and `.content` attributes set to `null` in order to support pre-compiled letter responses.
+
 ## [2.1.0] - 2018-08-14
 
 * The Notification class has a new `createdByName` property.